+++ site/pages/index.tsx
@@ -6,6 +6,7 @@ import { MalletExample } from "src/MalletExample";
 import { MellotronExample } from "src/MellotronExample";
 import { SmolkenExample } from "src/SmolkenExample";
 import { SoundfontExample } from "src/SoundfontExample";
+import { VersilianExample } from "src/VersilianExample";
 import { PianoExample } from "../src/PianoExample";
 
 const inter = Inter({ subsets: ["latin"] });
@@ -21,7 +22,7 @@ export default function Home() {
       <main className={"max-w-4xl mx-auto my-20 p-4" + inter.className}>
         <div className="flex items-end mb-16">
           <h1 className="text-6xl font-bold">smplr</h1>
-          <div>0.9.0</div>
+          <div>0.10.0</div>
         </div>
 
         <div className="flex flex-col gap-8">
@@ -32,6 +33,7 @@ export default function Home() {
           <DrumMachineExample />
           <MellotronExample />
           <SmolkenExample />
+          <VersilianExample />
         </div>
       </main>
     </>