+++ site/src/VersilianExample.tsx
@@ -1,0 +1,142 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import {
+  CacheStorage,
+  Reverb,
+  Storage,
+  Versilian,
+  getVersilianInstruments,
+} from "smplr";
+import { ConnectMidi } from "./ConnectMidi";
+import { PianoKeyboard } from "./PianoKeyboard";
+import { getAudioContext } from "./audio-context";
+import { LoadWithStatus, useStatus } from "./useStatus";
+
+let reverb: Reverb | undefined;
+let storage: Storage | undefined;
+let instrumentNames: string[] = [];
+
+// https://smpldsnds.github.io/sgossner-vcsl/Electrophones/TX81Z/FM%20Piano.sfz
+
+export function VersilianExample({ className }: { className?: string }) {
+  const [instrument, setInstrument] = useState<Versilian | undefined>(
+    undefined
+  );
+  const [instrumentName, setInstrumentName] = useState<string>(
+    instrumentNames[0]
+  );
+  const [status, setStatus] = useStatus();
+  const [reverbMix, setReverbMix] = useState(0);
+  const [volume, setVolume] = useState(100);
+
+  useEffect(() => {
+    getVersilianInstruments().then((names) => {
+      instrumentNames = names;
+      setInstrumentName(names[0]);
+    });
+  }, []);
+
+  function loadVersilian(instrumentName: string) {
+    if (instrument) instrument.disconnect();
+    setStatus("loading");
+    const context = getAudioContext();
+    reverb ??= new Reverb(context);
+    storage ??= new CacheStorage("smolken");
+    const newInstrument = new Versilian(context, {
+      instrument: instrumentName,
+      volume,
+    });
+    newInstrument.output.addEffect("reverb", reverb, reverbMix);
+    setInstrument(newInstrument);
+    newInstrument.load.then(() => {
+      setStatus("ready");
+    });
+  }
+
+  return (
+    <div className={className}>
+      <div className="flex gap-2 items-end mb-2">
+        <div className="flex flex-col">
+          <h1 className="text-3xl">Versilian </h1>
+          <h2>Versilian Community Sample Library</h2>
+        </div>
+
+        <div>(experimental)</div>
+
+        <LoadWithStatus
+          status={status}
+          onClick={() => loadVersilian(instrumentName)}
+        />
+        <ConnectMidi instrument={instrument} />
+      </div>
+      <div></div>
+      <div className={status !== "ready" ? "opacity-30" : ""}>
+        <div className="flex gap-4 mb-2 no-select">
+          <select
+            className="appearance-none bg-zinc-700 text-zinc-200 rounded border border-gray-400 py-2 px-3 leading-tight focus:outline-none focus:border-blue-500"
+            value={instrumentName}
+            onChange={(e) => {
+              const instrumentName = e.target.value;
+              loadVersilian(instrumentName);
+              setInstrumentName(instrumentName);
+            }}
+          >
+            {instrumentNames.map((name) => (
+              <option key={name} value={name}>
+                {name}
+              </option>
+            ))}
+          </select>
+          <button
+            className="bg-zinc-700 rounded px-3 py-0.5 shadow"
+            onClick={() => {
+              instrument?.stop();
+            }}
+          >
+            Stop all
+          </button>
+        </div>
+        <div className="flex gap-4 mb-2 no-select">
+          <div>Volume:</div>
+          <input
+            type="range"
+            min={0}
+            max={127}
+            step={1}
+            value={volume}
+            onChange={(e) => {
+              const volume = e.target.valueAsNumber;
+              instrument?.output.setVolume(volume);
+              setVolume(volume);
+            }}
+          />
+          <div>Reverb:</div>
+          <input
+            type="range"
+            min={0}
+            max={1}
+            step={0.001}
+            value={reverbMix}
+            onChange={(e) => {
+              const mix = e.target.valueAsNumber;
+              instrument?.output.sendEffect("reverb", mix);
+              setReverbMix(mix);
+            }}
+          />
+        </div>
+        <PianoKeyboard
+          borderColor="border-rose-700"
+          onPress={(note) => {
+            if (!instrument) return;
+            note.time = (note.time ?? 0) + instrument.context.currentTime;
+            instrument.start(note);
+          }}
+          onRelease={(midi) => {
+            instrument?.stop({ stopId: midi });
+          }}
+        />
+      </div>
+    </div>
+  );
+}