+++ src/player/types.ts
@@ -27,6 +27,8 @@ export type SampleOptions = {
   loop?: boolean;
   loopStart?: number;
   loopEnd?: number;
+  // not implemented yet
+  gainOffset?: number;
 };
 
 export type SampleStart = {
@@ -74,15 +76,24 @@ export type SampleRegion = {
   bendUp?: number;
   bendDown?: number;
 
-  offsetVol?: number;
-  offsetDetune?: number;
-
   /**
    * Velocity-based amplitude scaling. [Vel, Gain] tells the sampler to play the sample
    * at volume Gain when the note's velocity is Vel
    */
   ampVelCurve?: [number, number];
-  sample?: Partial<SampleOptions>;
+  /**
+   * Amplitude envelope release time in seconds.
+   * Stored here for convenience (flatness) but needs to be
+   * copied inside sample options before playback
+   */
+  ampRelease?: number;
+
+  /**
+   * Attack time in seconds. Currently not implemented
+   *
+   * @see http://sfzformat.com/opcodes/amp_attack.html
+   */
+  ampAttack?: number;
 
   /**
    * seqLength defines how many samples are in the sequence.
@@ -98,13 +109,6 @@ export type SampleRegion = {
   seqPosition?: number;
 
   /**
-   * Amplitude envelope release time in seconds.
-   * Stored here for convenience (flatness) but needs to be
-   * copied inside sample options before playback
-   */
-  ampRelease?: number;
-
-  /**
    * This assigns the group to a specific number.
    * Group numbers can be used in combination with groupOffBy to implement
    * exclusive groups, where playing one sample can stop another sample from playing.
@@ -115,9 +119,31 @@ export type SampleRegion = {
    * currently playing in the specified group number
    */
   groupOffBy?: number;
+
+  /**
+   * Start offset (in samples). Not implemented (yet)
+   */
+  offset?: number;
+
+  /**
+   * Adjust the playback pitch of a sample (in semitones)
+   */
+  tune?: number;
+
+  /**
+   * The volume opcode in SFZ defines the default playback volume for a given region.
+   * It specifies an adjustment to the sample's original amplitude.
+   * The unit for the volume opcode is decibels (dB).
+   */
+  volume?: number;
+
+  /**
+   * sample options for this particular region
+   */
+  sample?: Partial<SampleOptions>;
 };
 
-export type SampleLayer = {
+export type RegionGroup = {
   regions: SampleRegion[];
   sample: Partial<SampleOptions>;
 };