+++ src/smolken.ts
@@ -1,20 +1,16 @@
 import { ChannelOptions } from "./player/channel";
 import { DefaultPlayer } from "./player/default-player";
-import { createEmptySampleLayer, findSamplesInLayer } from "./player/layers";
-import {
-  AudioBuffers,
-  getPreferredAudioExtension,
-  loadAudioBuffer,
-} from "./player/load-audio";
+import { createEmptyRegionGroup, findSamplesInRegions } from "./player/layers";
+import { AudioBuffers } from "./player/load-audio";
 import { toMidi } from "./player/midi";
 import {
   InternalPlayer,
-  SampleLayer,
+  RegionGroup,
   SampleOptions,
   SampleStart,
   SampleStop,
 } from "./player/types";
-import { sfzToLayer } from "./sfz2";
+import { SfzInstrumentLoader } from "./sfz2";
 import { HttpStorage, Storage } from "./storage";
 
 export function getSmolkenNames() {
@@ -40,7 +36,7 @@ export type SmolkenOptions = Partial<
 
 export class Smolken implements InternalPlayer {
   private readonly player: DefaultPlayer;
-  private readonly layer: SampleLayer;
+  private readonly group: RegionGroup;
   public readonly load: Promise<this>;
   private config: SmolkenConfig;
   private seqNum = 0;
@@ -51,13 +47,19 @@ export class Smolken implements InternalPlayer {
       storage: options.storage ?? HttpStorage,
     };
     this.player = new DefaultPlayer(context, options);
-    this.layer = createEmptySampleLayer();
+    this.group = createEmptyRegionGroup();
+    const url = getSmolkenUrl(this.config.instrument);
 
-    const loader = loadSmolkenInstrument(
-      this.config.instrument,
-      this.player.buffers,
-      this.layer
-    );
+    const loader = SfzInstrumentLoader(url, {
+      buffers: this.player.buffers,
+      group: this.group,
+      urlFromSampleName: (sampleName, audioExt) => {
+        const samplePath = sampleName
+          .replace("\\", "/")
+          .replace(".wav", audioExt);
+        return `https://smpldsnds.github.io/sfzinstruments-dsmolken-double-bass/${samplePath}`;
+      },
+    });
     this.load = loader(context, this.config.storage).then(() => this);
   }
 
@@ -74,8 +76,8 @@ export class Smolken implements InternalPlayer {
   }
 
   start(sample: SampleStart | string | number) {
-    const found = findSamplesInLayer(
-      this.layer,
+    const found = findSamplesInRegions(
+      this.group,
       typeof sample === "object" ? sample : { note: sample },
       this.seqNum
     );
@@ -100,35 +102,4 @@ export class Smolken implements InternalPlayer {
   disconnect(): void {
     this.player.disconnect();
   }
-}
-
-function loadSmolkenInstrument(
-  instrumentOrUrl: string,
-  buffers: AudioBuffers,
-  layer: SampleLayer
-) {
-  const url = instrumentOrUrl.startsWith("http")
-    ? instrumentOrUrl
-    : getSmolkenUrl(instrumentOrUrl);
-  const audioExt = getPreferredAudioExtension();
-
-  return async (context: BaseAudioContext, storage: Storage) => {
-    const sfz = await fetch(url).then((res) => res.text());
-    const errors = sfzToLayer(sfz, layer);
-    if (errors.length) {
-      console.warn("Problems converting sfz", errors);
-    }
-    const sampleNames = new Set<string>();
-    layer.regions.forEach((r) => sampleNames.add(r.sampleName));
-    return Promise.all(
-      Array.from(sampleNames).map(async (sampleName) => {
-        const samplePath = sampleName
-          .replace("\\", "/")
-          .replace(".wav", audioExt);
-        const sampleUrl = `https://smpldsnds.github.io/sfzinstruments-dsmolken-double-bass/${samplePath}`;
-        const buffer = await loadAudioBuffer(context, sampleUrl, storage);
-        buffers[sampleName] = buffer;
-      })
-    );
-  };
 }