+++ src/soundfont/soundfont-instrument.ts
@@ -1,7 +1,7 @@
 import { spreadRegions } from "../player/layers";
 import { AudioBuffers, findFirstSupportedFormat } from "../player/load-audio";
 import { toMidi } from "../player/midi";
-import { SampleLayer } from "../player/types";
+import { RegionGroup } from "../player/types";
 import { Storage } from "../storage";
 
 export function gleitzKitUrl(name: string, kit: string) {
@@ -12,7 +12,7 @@ export function gleitzKitUrl(name: string, kit: string) {
 export function soundfontInstrumentLoader(
   url: string,
   buffers: AudioBuffers,
-  layer: SampleLayer
+  group: RegionGroup
 ) {
   return async (context: BaseAudioContext, storage: Storage) => {
     const sourceFile = await (await storage.fetch(url)).text();
@@ -28,13 +28,13 @@ export function soundfontInstrumentLoader(
         );
         const buffer = await context.decodeAudioData(audioData);
         buffers[noteName] = buffer;
-        layer.regions.push({
+        group.regions.push({
           sampleName: noteName,
           midiPitch: midi,
         });
       })
     );
-    spreadRegions(layer.regions);
+    spreadRegions(group.regions);
   };
 }